LayoutTests/platform/wpe/imported/w3c/web-platform-tests/mixed-content/gen/top.meta/unset/sharedworker-import-data.https-expected.txt


PASS Mixed-Content: Expects allowed for sharedworker-import-data to same-https origin and keep-scheme redirection from https context.
PASS Mixed-Content: Expects allowed for sharedworker-import-data to same-https origin and no-redirect redirection from https context.
FAIL Mixed-Content: Expects blocked for sharedworker-import-data to cross-http origin and keep-scheme redirection from https context. assert_equals: The resource request should be 'blocked'. expected "blocked" but got "allowed"
FAIL Mixed-Content: Expects blocked for sharedworker-import-data to cross-http origin and no-redirect redirection from https context. assert_equals: The resource request should be 'blocked'. expected "blocked" but got "allowed"
FAIL Mixed-Content: Expects blocked for sharedworker-import-data to cross-http origin and swap-scheme redirection from https context. assert_equals: The resource request should be 'blocked'. expected "blocked" but got "allowed"
FAIL Mixed-Content: Expects blocked for sharedworker-import-data to same-http origin and keep-scheme redirection from https context. assert_equals: The resource request should be 'blocked'. expected "blocked" but got "allowed"
FAIL Mixed-Content: Expects blocked for sharedworker-import-data to same-http origin and no-redirect redirection from https context. assert_equals: The resource request should be 'blocked'. expected "blocked" but got "allowed"
FAIL Mixed-Content: Expects blocked for sharedworker-import-data to same-http origin and swap-scheme redirection from https context. assert_equals: The resource request should be 'blocked'. expected "blocked" but got "allowed"
